Rewriting tests
