+++ app/test/app_test.go
@@ -1,13 +1,8 @@
-package app_test
+package test
 
 import (
-	"crypto/ecdsa"
-	"encoding/json"
-	"errors"
 	"io/ioutil"
-	"math/big"
 	"os"
-	"path/filepath"
 	"testing"
 	"time"
 
@@ -15,51 +10,17 @@ import (
 
 	"golang.org/x/net/context"
 
-	ethUtils "github.com/ethereum/go-ethereum/cmd/utils"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core"
-	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/ethereum/go-ethereum/eth"
-	"github.com/ethereum/go-ethereum/node"
 	"github.com/ethereum/go-ethereum/rlp"
 
-	"github.com/tendermint/ethermint/app"
-	emtUtils "github.com/tendermint/ethermint/cmd/utils"
-	"github.com/tendermint/ethermint/ethereum"
-
 	abciTypes "github.com/tendermint/abci/types"
-
-	ctypes "github.com/tendermint/tendermint/rpc/core/types"
 )
 
 var (
 	receiverAddress = common.StringToAddress("0x1234123412341234123412341234123412341234")
 )
-
-// implements: tendermint.rpc.client.HTTPClient
-type MockClient struct {
-	sentBroadcastTx chan struct{} // fires when we call broadcast_tx_sync
-}
-
-func NewMockClient() *MockClient { return &MockClient{make(chan struct{})} }
-
-func (mc *MockClient) Call(method string, params map[string]interface{}, result interface{}) (interface{}, error) {
-	_ = result
-	switch method {
-	case "status":
-		result = &ctypes.ResultStatus{}
-
-		return result, nil
-	case "broadcast_tx_sync":
-		close(mc.sentBroadcastTx)
-		result = &ctypes.ResultBroadcastTx{}
-
-		return result, nil
-	}
-
-	return nil, abciTypes.ErrInternalError
-}
 
 func TestBumpingNonces(t *testing.T) {
 	// generate key
@@ -267,88 +228,4 @@ func TestMultipleTxTwoAcc(t *testing.T) {
 	assert.Equal(t, abciTypes.OK.Code, app.Commit().Code)
 
 	node.Stop() // nolint: errcheck
-}
-
-// mimics abciEthereumAction from cmd/ethermint/main.go
-func makeTestApp(tempDatadir string, addresses []common.Address, mockclient *MockClient) (*node.Node, *ethereum.Backend, *app.EthermintApplication, error) {
-	stack, err := makeTestSystemNode(tempDatadir, addresses, mockclient)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	ethUtils.StartNode(stack)
-
-	var backend *ethereum.Backend
-	if err = stack.Service(&backend); err != nil {
-		return nil, nil, nil, err
-	}
-
-	app, err := app.NewEthermintApplication(backend, nil, nil)
-
-	return stack, backend, app, err
-}
-
-func makeTestGenesis(addresses []common.Address) (*core.Genesis, error) {
-	gopath := os.Getenv("GOPATH")
-	genesisPath := filepath.Join(gopath, "src/github.com/tendermint/ethermint/setup/genesis.json")
-
-	file, err := os.Open(genesisPath)
-	if err != nil {
-		return nil, err
-	}
-	defer file.Close() // nolint: errcheck
-
-	genesis := new(core.Genesis)
-	if err := json.NewDecoder(file).Decode(genesis); err != nil {
-		ethUtils.Fatalf("invalid genesis file: %v", err)
-	}
-
-	balance, result := new(big.Int).SetString("10000000000000000000000000000000000", 10)
-	if !result {
-		return nil, errors.New("BigInt convertation error")
-	}
-
-	for _, addr := range addresses {
-		genesis.Alloc[addr] = core.GenesisAccount{Balance: balance}
-	}
-
-	return genesis, nil
-}
-
-// mimics MakeSystemNode from ethereum/node.go
-func makeTestSystemNode(tempDatadir string, addresses []common.Address, mockclient *MockClient) (*node.Node, error) {
-	// Configure the node's service container
-	nodeConf := emtUtils.DefaultNodeConfig()
-	emtUtils.SetEthermintNodeConfig(&nodeConf)
-	nodeConf.DataDir = tempDatadir
-
-	// Configure the Ethereum service
-	ethConf := eth.DefaultConfig
-	emtUtils.SetEthermintEthConfig(&ethConf)
-
-	genesis, err := makeTestGenesis(addresses)
-	if err != nil {
-		return nil, err
-	}
-
-	ethConf.Genesis = genesis
-
-	// Assemble and return the protocol stack
-	stack, err := node.New(&nodeConf)
-	if err != nil {
-		return nil, err
-	}
-	return stack, stack.Register(func(ctx *node.ServiceContext) (node.Service, error) {
-		return ethereum.NewBackend(ctx, &ethConf, mockclient)
-	})
-}
-
-func createTransaction(key *ecdsa.PrivateKey, nonce uint64) (*types.Transaction, error) {
-	signer := types.HomesteadSigner{}
-
-	return types.SignTx(
-		types.NewTransaction(nonce, receiverAddress, big.NewInt(10), big.NewInt(21000), big.NewInt(10),
-			nil),
-		signer,
-		key,
-	)
 }
